requirements.txt
# Web framework
fastapi==0.129.*
uvicorn[standard]==0.41.*
jinja2==3.1.*

# Database
sqlalchemy==2.0.*
psycopg2-binary==2.9.*

# HTTP client (Perplexity API calls)
httpx==0.28.*

# Settings
pydantic-settings==2.13.*

# Testing
pytest==9.0.*
pytest-httpx==0.36.*
